enterprise-apps/crm-systems/lead-management/requirements.txt
# Django
Django==4.2.9
django-environ==0.11.2

# Database
mysqlclient==2.2.1

# REST API
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5

# Forms
django-crispy-forms==2.1
crispy-bootstrap5==2.0.0

# Background tasks
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
Pillow==10.1.0

# Import/Export
django-import-export==3.3.5
openpyxl==3.1.2

# Development
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Testing
pytest==7.4.4
pytest-django==4.7.0
coverage==7.4.0
factory-boy==3.3.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
